Fix Markdown Engine Error
Current compilation of the Developer's Blog throws an error during compilation because we were using a deprecated syntax highlighter and markdown engine. This fixes both errors and updates us to using `kramdown` markdown enigne and the `rouge` syntax highlighter. Also fixes some issues with a post related to syntax highlighting